Create gameLinker returnsSingleGameId.html
A very rough one at a time working concept!  Currently only works if the name in the JSON is the same as the input (which is ok because most game names copied from humble bundle seem to work fine).
It's kind of a sloppy mess and sorry if you're looking at it.  Eventually I will split the javascript away from the tiny bit of html.